changed constants to CODATA 2006; device exps and coef are in double precision

+++ cuda_rys_dp.h
@@ -49,8 +49,8 @@ __device__ double cuda_rys_coulomb_repulsion_dp(double xa,double ya,double za,do
                                                 int lc,int mc,int nc,double alphac,
                                                 double xd,double yd,double zd,double normd,
                                                 int ld,int md,int nd,double alphad);
-__global__ void cuda_rys_eri_2d_dp(double *xa,double *ya,double *za,
-                                   int *la,int *ma,int *na,float *aexps,float *acoef,
-                                   double *xb,double *yb,double *zb,
-                                   int *lb,int *mb,int *nb,float *bexps,float *bcoef,
+__global__ void cuda_rys_eri_2d_dp(double *xa, double *ya, double *za, 
+                                   int *la, int *ma, int *na, double *aexps, double *acoef, 
+                                   double *xb, double *yb, double *zb, 
+                                   int *lb, int *mb, int *nb, double *bexps, double *bcoef, 
                                    int n_combi, int *start_contr, int *end_contr, double *eri);
